Use location.replace for URL changes to fix back-button problems.
By using location.replace instead of location.hash to update the URL a
new history entry is not created, so the back button will now behave
as the user expects an return to the page before the bookreader.

This should work for IE and Firefox, but apparantly may not for WebKit
based browsers, as location.replace behaves differently for them.  See
these reported WebKit bugs:

https://bugs.webkit.org/show_bug.cgi?id=20355
https://bugs.webkit.org/show_bug.cgi?id=20425
https://bugs.webkit.org/show_bug.cgi?id=22224

This fix (at least partially) addresses these GnuBook bugs:

https://bugs.launchpad.net/gnubook/+bug/300521
https://bugs.launchpad.net/gnubook/+bug/294281
https://bugs.launchpad.net/gnubook/+bug/272022