Implement true fullscreen
When the user presses the "toggle fullscreen" button, the BookReader
expands to fill the whole screen, using the browser's built-in
Fullscreen API (as used by YouTube, for example). The user gets a
"true" fullscreen experience, with no distractions.

When the user presses the ESC button, or the "toggle fullscreen"
button again, then the fullscreen mode exits.

This patch also ensures that the "share", "info" and "download"
popup modals are properly displayed in fullscreen mode.

If fullscreen mode is not supported for a given device or document,
then gracefully falls back to the legacy "full window" behavior.