WEBDEV-2121: Fix Rapid Reload Bug in archive.org `/stream` View
There is a bug in the archive.org `/stream` view where after a user selects a page from the thumbnail view, the page starts rapidly switching between the thumbnail view and the single page view.

It was being caused by a double call to `trigger(BookReader.eventNames.fragmentChange)`, one in `updateFirstIndex()` and another in `switchMode()`.

This adds a parameter to suppress the `trigger` calls so we can make both of the calls fist, then call `trigger` when they're done.