WEBDEV-1992: Fix Rapid Reload Bug from Thumbnail View (#102)
### Background
There is a bug in the archive.org `/stream` view where after a user selects a page from the thumbnail view, the page starts rapidly switching between the thumbnail view and the full page view, causing the browser to freeze.

### Cause
There were two calls being made `trigger(BookReader.eventNames.fragmentChange)`, one in `updateFirstIndex()` and another in `switchMode()`. This was triggering other listeners that kept calling back into `switchMode()` and we'd get caught in the loop.

### Solution
This adds a parameter to suppress the `trigger` calls so we can make both of the calls first, then call `trigger` when they're done.